Update README.md
There are a lot of dustbins in a
shopping mall and it is a hassle for the
cleaners to check all of them one by one
whether it is full. Therefore, a smart dustbin
with rubbish fullness detection is introduced.
Ultrasonic sensor is chosen to be the sensor to
detect the distance of the rubbish from the
cap of the dustbin to determine the level of
the rubbish in the dustbin. Arduino is used
for designing and programming the circuit.
For Graphical User Interface, LED indicator
and LCD is used to display the fullness of the
rubbish in the dustbin to the cleaner. The
characteristic of the sensor is determined by
using different materials to test the detection
of the ultrasonic sensor.